# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Saina Gallery RESTful API",
        "description": "SainaGallery API documentation"
    },
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/gallery": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get one Gallery",
                "description": "Get one Gallery",
                "operationId": "getonegallery",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id",
                        "description": "Gallery id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/seller/product": {
            "post": {
                "tags": [
                    "seller"
                ],
                "summary": "Add New Product",
                "description": "Add Product",
                "operationId": "newproduct",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "productImage",
                        "type": "file",
                        "in": "formData"
                    },
                    {
                        "name": "name",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "category",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "description",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "number",
                        "type": "number",
                        "in": "formData"
                    },
                    {
                        "name": "price",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "isDiscounted",
                        "type": "boolean",
                        "in": "formData"
                    },
                    {
                        "name": "priceWithDiscount",
                        "type": "string",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "seller"
                ],
                "summary": "Edit a product",
                "description": "Edit product",
                "operationId": "editproduct",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "type": "string"
                    },
                    {
                        "name": "productImage",
                        "type": "file",
                        "in": "formData"
                    },
                    {
                        "name": "category",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "name",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "description",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "number",
                        "type": "number",
                        "in": "formData"
                    },
                    {
                        "name": "price",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "isDiscounted",
                        "type": "boolean",
                        "in": "formData"
                    },
                    {
                        "name": "priceWithDiscount",
                        "type": "string",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "seller"
                ],
                "summary": "Delete one product",
                "description": "Delete one product",
                "operationId": "deleteproduct",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Delete one Province by id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/user": {
            "get": {
                "tags": [
                    "api"
                ],
                "summary": "Get One User",
                "description": "Get One User",
                "operationId": "getoneuser",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/api/seller/changefactorcondition": {
            "get": {
                "tags": [
                    "seller"
                ],
                "summary": "change factor condition",
                "description": " change factor condition",
                "operationId": "changefactorcondition",
                "parameters": [
                    {
                        "in": "query",
                        "type": "string",
                        "name": "id"
                    },
                    {
                        "in": "query",
                        "type": "string",
                        "name": "condition"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/seller/changeadsstatus": {
            "get": {
                "tags": [
                    "seller"
                ],
                "summary": "change ads status",
                "description": " change ads status",
                "operationId": "changeadsstatus",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
...
}